+++ crates/bevy_flurx_api/src/web_window/create.rs
@@ -6,7 +6,7 @@ use bevy::window::{Window, WindowResolution};
 use bevy_flurx::action::once;
 use bevy_flurx::prelude::Action;
 use bevy_flurx_ipc::prelude::*;
-use bevy_webview_wry::prelude::{AutoPlay, BrowserAcceleratorKeys, HotkeysZoom, Incognito, InitializeFocused, IsOpenDevtools, Theme, UseDevtools, UseHttpsScheme, UserAgent, Webview, WebviewUri, WebviewVisible};
+use bevy_webview_core::prelude::*;
 use serde::Deserialize;
 use winit::dpi::PhysicalSize;