+++ crates/bevy_flurx_ipc/src/component.rs
@@ -46,8 +46,8 @@ impl IpcHandlers {
     }
 
     /// Returns the function that creates the future if exists related to `id`.
+    #[inline]
     pub fn get(&self, id: &str) -> Option<IpcFn> {
-        println!("IpcHandlers::get : id = {} handlers = {:?}", id, self.0.keys());
         self.0.get(id).map(|handler| handler.f)
     }
 }
@@ -88,6 +88,7 @@ impl IpcHandler {
     }
 
     /// Returns the ipc-id.
+    #[inline]
     pub fn id(&self) -> &str {
         &self.id
     }