+++ crates/bevy_flurx_ipc/src/ipc_commands.rs
@@ -91,7 +91,7 @@ pub struct IpcResolveEvent {
     pub output: String,
 }
 
-/// The common plugin for IPC communication between `Webview` and `bevy`.
+/// The common webview for IPC communication between `Webview` and `bevy`.
 pub(crate) struct FlurxIpcCommandPlugin;
 
 impl Plugin for FlurxIpcCommandPlugin {
@@ -110,7 +110,6 @@ fn receive_ipc_commands(
     handlers: Query<&IpcHandlers>,
 ) {
     for cmd in ipc_commands.take_commands() {
-        println!("receive_ipc_commands: {cmd:?}");
         let Ok(handlers) = handlers.get(cmd.entity) else {
             continue;
         };