+++ crates/bevy_webview_core/src/lib.rs
@@ -1,0 +1,11 @@
+//! Defines the core webview logics.
+//!
+//! This crate is a part of [`bevy_webview_projects`](https://github.com/not-elm/bevy_webview_projects).
+
+/// Defines the core webview components.
+pub mod bundle;
+
+#[allow(missing_docs)]
+pub mod prelude {
+    pub use crate::bundle::prelude::*;
+}