+++ crates/bevy_webview_wry/src/lib.rs
@@ -1,25 +1,41 @@
 //! Provides the minimum functionality required to display webview.
 
-use crate::common::plugin::FlurxWryCommonPlugin;
-use crate::embedding::plugin::AsChildPlugin;
+use crate::embedding::EmbeddingWebviewPlugin;
 use bevy::prelude::{App, Plugin, Reflect, Resource};
+use bevy_webview_core::bundle::WebViewCoreBundlesPlugin;
 use std::path::PathBuf;
+use webview::WebviewPlugin;
 
-#[allow(missing_docs)]
+/// [`bevy_webview_core`]
+pub mod core {
+    pub use bevy_webview_core::prelude::*;
+}
+
+/// [`bevy_flurx_ipc`]
 pub mod ipc {
     pub use bevy_flurx_ipc::prelude::*;
 }
 
+
+#[cfg(feature = "api")]
+/// [`bevy_flurx_api`]
+pub mod api {
+    pub use bevy_flurx_api::prelude::*;
+}
+
 pub mod embedding;
-pub mod common;
+pub mod webview;
 mod util;
 
 #[allow(missing_docs)]
 pub mod prelude {
-    pub use crate::{common::prelude::*, embedding::prelude::*, WebviewWryPlugin};
+    pub use crate::{embedding::prelude::*, webview::prelude::*, WebviewWryPlugin};
     #[cfg(feature = "child_window")]
     pub use bevy_child_window::prelude::*;
-    pub use bevy_flurx_ipc::prelude::command;
+    #[cfg(feature = "api")]
+    pub use bevy_flurx_api::prelude::*;
+    pub use bevy_flurx_ipc::prelude::*;
+    pub use bevy_webview_core::prelude::*;
 }
 
 #[repr(transparent)]
@@ -30,7 +46,7 @@ pub(crate) struct WryLocalRoot(pub PathBuf);
 /// in a [`Window`](bevy::prelude::Window) using [`wry`].
 pub struct WebviewWryPlugin {
     /// Represents the root directory of the local resource.
-    /// This value affects [`WebviewUri`](crate::prelude::WebviewUri).
+    /// This value affects [`WebviewUri`](prelude::WebviewUri).
     ///
     /// This directory must be located under the `assets` directory.
     pub local_root: PathBuf,
@@ -49,7 +65,15 @@ impl Plugin for WebviewWryPlugin {
         app
             .register_type::<WryLocalRoot>()
             .insert_resource(WryLocalRoot(self.local_root.clone()))
-            .add_plugins((FlurxWryCommonPlugin, AsChildPlugin));
+            .add_plugins((
+                WebviewPlugin,
+                EmbeddingWebviewPlugin,
+            ));
+
+        if !app.is_plugin_added::<WebViewCoreBundlesPlugin>() {
+            app.add_plugins(WebViewCoreBundlesPlugin);
+        }
+
         #[cfg(feature = "child_window")]
         {
             use bevy_child_window::ChildWindowPlugin;